removed folder mobile
