+++ src/reducers/index.js
@@ -1,0 +1,14 @@
+import { combineReducers } from "redux";
+import user from "./user";
+import newUser from "./newUser";
+import card from "./card";
+import order from "./order";
+import addresses from "./addresses";
+
+export default combineReducers({
+  user,
+  newUser,
+  order,
+  card,
+  addresses,
+});
