--- src/redux/index.js
@@ -1,4 +1,0 @@
-export * from "./actions";
-export * from "./middlewares";
-export * from "./selectors";
-export { default } from "./reducer";
