+++ src/selectors/selectors.js
@@ -16,9 +16,3 @@ export const getUser = createSelector(
     (state) => state.card,
     (card) => card
   );
-
-  
-// export const getUser = (state) => state.user;
-// export const getNewUser = (state) => state.newUser;
-// export const getOrder = (state) => state.order;
-// export const getCard = (state) => state.card;
